Add posshift compensator for line mode
