+++ test/index.js
@@ -1,5 +1,6 @@
 'use strict'
 
-require('./calibrate')
+document.body.style.margin = '0'
+
 require('./cases')
 require('./multipass')
